Another update to the update function
Ensure that Update.py is updated before any other Python files, as Update.py contains the names and locations of each .py module, which may change over time.

Subsequent push was to ensure that the main repo is only cloned once.

This push ensure that the Update function is reimported after it updates - I ddidn't realize that Python stores the intial version in memory. Because of this, there is now a dependency on importlib, which is now part of the checklist on GUI launch